Further cleanups, address reviews
