+++ twemproxy/tests/compose/docker-compose.yaml
@@ -29,19 +29,16 @@ services:
             -initial-cluster-token etcd-cluster-1
             -initial-cluster etcd0=http://0.0.0.0:2380
             -initial-cluster-state new
+    depends_on:
+      - redis1
+      - redis2
 
   twemproxy:
     image: jgoodall/twemproxy
     ports:
       - 6100:6100
       - 6222:6222
-    volumes:
-      - ${WAIT_FOR_IT_SCRIPT_PATH}:/wait-for-it.sh
-      - ${SETUP_SCRIPT_PATH}:/setup.sh
-    command: ["bash", "-c", "chmod +x /wait-for-it.sh && /wait-for-it.sh etcd0:4001 -t 120 -- bash /setup.sh"]
     environment:
       - ETCD_HOST=etcd0:4001
     depends_on:
-      - redis1
-      - redis2
       - etcd0
